Change store signature to support persistent stores (#50) (#60)
The original design of the store returned references to flags and
segments. This was safe to do because we knew the lifetime of the flag
or segment was scoped to the lifetime of the store and thus returning a
reference was safe.

With the introduction of persistent stores and the persistent wrapper,
we can no longer return references. The SDK's cache hands back full
objects, and since the borrow checker will not allow us to return
references to function scoped values, we are left with little choice but
to modify the signature of this trait.