Resolve last few TODOs in repo (#112)
* Remove doc link from ContextAttributes

* Breaking change: remove ContextBuilder::set_int

* Remove TODO about setting name via ContextBuilder::try_set_value

The question was if it should allow setting 'name' to AttributeValue::Null, which would
unset it. I've adopted this behavior, and removed the TODO note.

The justification is that 'name' is an optional property, and so there should be a way of unsetting it (not possible with
the .name() setter).